==> 06 Pipeline/P1 titanic_without_pipline.ipynb <==
<h1 align = 'center'> Titanic witout using Pipeline </h1>

- Piplines chains together multiple steps so that ouput of each step is used as input to the next step.
- Pipline makes it easy to apply same preprocessing to train and test!

import numpy as np
import pandas as pd

df = pd.read_csv('titanic_train.csv')

df.head()

  df.columns

#Droping unwanted columns
df.drop(columns=['PassengerId','Name','Ticket','Cabin'], inplace=True)

df.head()

### Train Test Split

from sklearn.model_selection import train_test_split

X = df.drop('Survived', axis=1)

X

y= df['Survived']

y

X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.2, random_state=42)

X_train.head(2)

y_train.head(2)

### Missing values

df.isnull().sum()

from sklearn.impute import SimpleImputer

### Apply Imputation to -> Age(Num), Embarked(Categorical)
si_age= SimpleImputer() ## Filling missing value using default statergy:Mean 
si_embarked = SimpleImputer(strategy='most_frequent')# Filling missing value using Most_frequent value

X_train_age = si_age.fit_transform(X_train[['Age']])
X_train_embarked = si_embarked.fit_transform(X_train[['Embarked']])

X_test_age =si_age.transform(X_test[['Age']])
X_test_embarked =si_embarked.transform(X_test[['Embarked']])



X_train_embarked

### One Hot Encoding

- Categorical -- Sex, Embarked 
- One hot Encoding

- Now we will create 2 object for Sex, Embarked 
- Beccause eventhough we have solved and created new array X_train_embarked with filling value, the actual X_train still have missing value on wich we will apply transform

from sklearn.preprocessing import OneHotEncoder

# one hot encoding Sex and Embarked

ohe_sex = OneHotEncoder(sparse=False,handle_unknown='ignore')
ohe_embarked = OneHotEncoder(sparse=False,handle_unknown='ignore')

X_train_sex = ohe_sex.fit_transform(X_train[['Sex']])
X_train_embarked = ohe_embarked.fit_transform(X_train_embarked)

X_test_sex = ohe_sex.transform(X_test[['Sex']])
X_test_embarked = ohe_embarked.transform(X_test_embarked)

- handle_unkown= 'ignore' means in future we get completly new category in test data we will going to ignore it means it will fill with 0.
- Here we didn't drop the column for solving multicolinearity as we will going to apply DT to model it dosn't have much effect

X_train_embarked

- Now we have three array X_train_age, X_train_sex, X_train_embarked 
- All other columns made no change
- Now we concatinate it to the all other array


X_train_rem = X_train.drop(columns=['Sex','Age','Embarked'])

X_test_rem = X_test.drop(columns=['Sex','Age','Embarked'])

X_train_transform = np.concatenate((X_train_rem, X_train_age, X_train_sex, X_train_embarked), axis=1)
X_test_transform = np.concatenate((X_test_rem, X_test_age, X_test_sex, X_test_embarked), axis=1)

### Apply it to  Decision Tree

from sklearn.tree import DecisionTreeClassifier

clf = DecisionTreeClassifier()

clf.fit(X_train_transform,y_train)

y_pred = clf.predict(X_test_transform) 

y_pred

### Performance parameter of DT

from sklearn.metrics import accuracy_score

accuracy_score(y_test,y_pred)

# Export 

import pickle

pickle.dump(ohe_sex, open('D:\Python\statistics\ohe_sex.pkl', 'wb'))
pickle.dump(ohe_embarked, open('D:\Python\statistics\ohe_embarked.pkl','wb'))
pickle.dump(clf, open('D:\Python\statistics\clf.pkl','wb'))
