==> 06 Pipeline/P2 predict_without_pipline.ipynb <==
<h1 align='center'> Predict without pipline</h1>

import numpy as np
import pandas as pd
import pickle

### Import model

ohe_sex =pickle.load(open('D:\Python\statistics\ohe_sex.pkl', 'rb'))
ohe_embarked = pickle.load(open('D:\Python\statistics\ohe_embarked.pkl', 'rb'))
clf = pickle.load(open('D:\Python\statistics\clf.pkl','rb'))

### Assume  user Input

# Pclass/Sex/Age/SibSp/ParCh/Fare/Embarked
test_input = np.array([2,'male',31.0,0,0,10.5 ,'S'],dtype=object).reshape(1,7)

test_input

test_input_sex= ohe_sex.transform(test_input[:,1].reshape(1,1))

test_input_sex

test_input_embarked = ohe_embarked.transform(test_input[:,-1].reshape(1,1))

test_input_embarked

test_input_age = test_input[:,2].reshape(1,1)

- Here we created array for test input sex, age, Embarked now concatinate it to test input

# Skip age, sex, embarked values from test_input
test_input_transformed = np.concatenate((test_input[:,[0,3,4,5]],test_input_age, test_input_sex, test_input_embarked),axis=1)

test_input_transformed

clf.predict(test_input_transformed)
