==> 06 Pipeline/P3 titanic using Pipleine.ipynb <==
<h1 align='center'> Titanic using Pipline</h1>

import pandas as pd 
import numpy as np


df = pd.read_csv('titanic_train.csv')

df.head()

#Droping unwanted columns
df.drop(columns=['PassengerId','Name','Ticket','Cabin'], inplace=True)

 What is plan?
   1. Impute columns-- Age, Embarked
   2. Ohe -- Sex, Embarked
   3. Scaling
   4. feature Selection
   5. Model training using DT

### Train Test Split 

from sklearn.model_selection import train_test_split

X= df.drop(columns='Survived')

X

y=df['Survived']

X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.2, random_state=42)

X_train.head()

### Column Transformer

- Here we are not calling columns in class by there name but by index values as after transformer it will be convert into numpy array. Which will be given to next pipline 
- So it's Good strtatergy to apply index

from sklearn.compose import ColumnTransformer

### Simple inmputer : Age, Embarked

from sklearn.impute import SimpleImputer

trf1 = ColumnTransformer([
    ('impute_age', SimpleImputer(),[2]) ,# by Mean # [2] index of age
    ('impute_embarked',SimpleImputer(strategy='most_frequent'),[6]) # By most Frequent value in columns
], remainder='passthrough')

### One Hot Encoding: Sex, Embarkred

from sklearn.preprocessing import OneHotEncoder

# One hot encoder
trf2= ColumnTransformer([
    ('ohe_sex_embarked', OneHotEncoder(sparse=False,handle_unknown='ignore'),[1,6])
], remainder='passthrough')

- After appling above 2 transformer we get 10 column total

### Scaling :  All the Features

from sklearn.preprocessing import MinMaxScaler

# Saling
trf3 = ColumnTransformer([
    ('scaling',MinMaxScaler(), slice(0,10)) # All Column
])

### Feature Selection

from sklearn.feature_selection import SelectKBest, chi2

- Here for Chi - Squared stats of non-negative features for classification task
- k =8 means top 8 features to be selected

trf4 =SelectKBest(score_func= chi2, k=5)

### Model Training : DT

from sklearn.tree import DecisionTreeClassifier

trf5= DecisionTreeClassifier()

### Pipline

from sklearn.pipeline import Pipeline, make_pipeline

pipe =Pipeline([
    ('trf1',trf1), # Name and transformer object
    ('trf2',trf2),
    ('trf3',trf3),
    ('trf4',trf4),
    ('trf5',trf5)
])

### Pipline Vs make_pipline

- Pipline required nameing of steps, make_pipliine does not required
- (same applys To ColumnTransformer vs make_column_transformer)

#Alternative to pipline
pipe =make_pipeline(trf1,trf2,trf3,trf4,trf5)

### Train the model

- one thing sholud be noted that in pipline if we are only doing preprocessing i.e
- If we have only made pipline of SimpleImputer -> OneHotEncoder -> scaling , that time we would have apply __fit_transform__ 
- But in this pipline we did model training also that's why we will apply __fit__ to it and then __predict__

pipe.fit(X_train,y_train)

### To see Visual Representaion of pipline

## Display pipline
from sklearn import set_config

set_config(display='diagram')

pipe.fit(X_train,y_train)

## To see with transformer we use on the object

pipe.named_steps['trf1'].transformers_

##  If we want to find mean value we apply to impute_
##age
pipe.named_steps['trf1'].transformers_[0] 

- It gives tuple of trf1 to get simple imputer we will use index 1 which is SimpleImputer()

pipe.named_steps['trf1'].transformers_[0][1]

- SimpleImuter have attribute statistics  if we take it it will give us mean value

pipe.named_steps['trf1'].transformers_[0][1].statistics_

pipe.named_steps['trf1'].transformers_

- Now we want to See Most Frequent Value in impute_embarked

- 1 index Tuple of trf1, 1 index value which SimpleImputer object and apply statistics_

pipe.named_steps['trf1'].transformers_[1][1].statistics_

### Predict model

y_pred = pipe.predict(X_test)

y_pred

### Performance Metrics

from sklearn.metrics import accuracy_score

accuracy_score(y_test,y_pred)

- Here we use feature Selection where we reduce the features that's why accuracy is less
- So feature selction should be use wisely

### Crosss validation using pipline

from sklearn.model_selection import cross_val_score

cross_val_score(pipe, X_train, cv=5, scoring='accuracy').mean()

### Hyper Paramerter tuning using GridSearchCV

parm = {
    'trf5__max_depth':[1,2,3,4,5, None]
}

from sklearn.model_selection import GridSearchCV

grid = GridSearchCV(pipe, parm, cv=5, scoring='accuracy')

grid.fit(X_train, y_train)

grid.best_score_

grid.best_params_

### Exporting the Pipline

# Export 
import pickle

pickle.dump(pipe,open('pipe.pkl','wb'))
